Add components folder to separate UI components from other front end files
