+++ frontend/src/components/PreferencesCheckbox.js
