+++ frontend/src/components/ProductAreaDropdown.js
